- changed money type to BigDecimal from float - added new transactions to initial script

+++ src/main/java/io/stock/portfolio/backend/client/exchangerate/Rate.java
@@ -2,18 +2,20 @@ package io.stock.portfolio.backend.client.exchangerate;
 
 import com.fasterxml.jackson.annotation.JsonAlias;
 
+import java.math.BigDecimal;
+
 
 public class Rate {
 
-    private Float USD;
+    private BigDecimal USD;
 
     @JsonAlias("USD")
-    public Float getUSD() {
+    public BigDecimal getUSD() {
         return USD;
     }
 
     @JsonAlias("USD")
-    public void setUSD(Float USD) {
+    public void setUSD(BigDecimal USD) {
         this.USD = USD;
     }
