+++ src/main/java/io/stock/portfolio/backend/client/yahoo/YahooSplit.java
@@ -3,11 +3,12 @@ package io.stock.portfolio.backend.client.yahoo;
 import lombok.Data;
 import lombok.experimental.Accessors;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
 @Accessors(chain = true)
 public class YahooSplit {
     private LocalDateTime date;
-    private Float multiplier;
+    private BigDecimal multiplier;
 }
